Added `stitch` package.
This package works around the issue that we cannot use shared visualization infrastructure for
both local and cloud environments.

Recent changes made around some of the cloud environments allow us to use a jury-rigged IPyWidget as
a bidirectional communication channel between the kernel and frontend.
We allow the user to run their own HTML/JS but from within a sandboxed iframe that communicates to the
kernel first via postMessage to the widget's frontend, and onwards to the kernel.