Initial commit for benchmarking platform: `powerlift`.
Powerlift is a benchmarking platform for machine learning that prioritizes scale of datasets as well as quick, continual feedback to the user.
It is a spiritual successor to what is used internally within InterpretML for benchmarking EBMs against other methods.
This is not fully complete - there still needs to be easy access to experiment results as well as connected visualizations (via stitch).